Java HIDL interface libraries don't include hwbinder.jar
hwbinder.jar is built withint SDK (i.e. it uses platform internal APIs).
The library is statically embedded in Java HIDL interface libraries,
which in turn statically linked to apps. This makes the app binary to
have references to the internal APIs, which is wrong and will be
reported as link-type check errors when the check routine is run even
for Soong-to-Soong dependencies. (b/73829976)

Fixing the problem by removing hwbinder.jar from the static dependency.
The symbols that have been provided by hwbinder.jar are provided via the
SystemSDK stubs (sdk_version:"system_current"), so the dependency to
hwbinder.jar is in fact not needed.

Bug: 73829976
Test: m -j
Change-Id: I6a7b13262a91dc8c0d2a54dc0ab2ecd2de30fcdd