FQName: prepare for removing isValid
Having an object which has an invalid state
was a bad idea to begin with. We want this data
in the type system (whenever you get an FQName).

This CL adds a parse method to FQName which will
become the only way to create an FQName. The follow-up
for this CL will be to update all users of the CL
and then remove isValid.

Bug: 73774955
Test: hidl's run_all_host_tests.sh
Change-Id: I9a020dae27f7c7c2d0ce1bbb80830f28c3ea34e3