Add DT_UNKNOWN handling
Problem: With particular file system, d_type is not supported and the
value will be DT_UNKNOWN.

As readdir(3) states:
     Currently, only some filesystems (among them: Btrfs, ext2, ext3,
     and ext4) have full support for returning the file type in d_type.
     All applications must properly handle a return of DT_UNKNOWN.
Reference: http://man7.org/linux/man-pages/man3/readdir.3.html

Solution: If d_type is DT_UNKNOWN, use lstat(2) to check file mode.

Bug: 76443731

Test: With Linux kernel 4.16-rc5, overlayfs exported with NFSv3, d_type
is not supported. We verified the solution with above environment, and
build is passed with the patch. On the contrary, we get build
errors if without the solution.
Test: ./test/run_all_host_tests.sh (on ext4)

Change-Id: Ide175a301f7c4b8dc1106072fee1ff809e8fd5c4