Interchange union/discriminator positions in HIDL safe unions
This change swaps the positions of the union and discriminator in HIDL
safe_unions (new layout: discriminator, followed by the actual union).
This is necessary for unmarshalling certain types of parcels in Java
(ones containing interfaces), where the parcel contents are parsed
sequentially. In such cases, the discriminator value must be known
before the payload-specific functionality can be invoked.

Bug: 79878527
Test: Ran make and the hidl-test suite
Change-Id: Id65384681aef6bd72d4213073bf786674ed8cb22