safe_union: outside of android.hardware packages
This is a reoccuring issue (hidl-gen emitting things in the wrong
namespace even though they technically happen to be in the right
namespace for some builds. safe_union accidentally broke this (and
I've added a test for it), but it is the result of a tradeoff that
is being made. hidl-gen generates code by outputting text rather than
by outputting things into an intelligent C++ AST which can check these
kinds of things.

Since this is a relatively minor problem, for now, this is going to be
fixed in place again (with tests added).

Bug: N/A
Test: m hidl.tests.vendor@1.1 (checks the namespace at compile time)
Change-Id: Iee0ca1357c2c64d05598e18e27df8b7add5cb55d