safe_union: fix up default initialization.
This error was introduced when hidl_no_init was removed.

Before, Objecct hidl_o only got declared if a default Java initializer
was defined for this type. Now, this field is declared, and then the
default java value is set in the constructor.

This also means that primitive values must be manually initialized (or
they result in a NullPointerException). This is intentional, and we do
it for the same reason the Java wrapper classes don't have a default
value (so we don't initialize an extra object to a useless value).

In the future, we're still able to change this (it isn't part of any
ABI). One optimization to be made here is that we can add an extra
field to the safe union like "long hidl_o_primitive" which stores the
value of the fields if it is a primitive to avoid an object creation.
This would also mean at that point that primitives won't require an
additional object instantiation.

Bug: N/A
Test: hidl_test, hidl_test_java, manually inspect safe_union output.
Change-Id: I2317f7b72819de5132944ecc9c723203123fc55d