Add option to use current.txt.
The build system declaring this dependency much cleaner, and it is good
to start thinking about freezing modules when they are added.

Currently the default for this option is to not use current.txt. This
is being added so that if someone has a problem with this file and
needs to use it, they can specify this option, and if the file is in
the wrong place, they'll get an appropriate error message. In the
future, if this isn't set, hidl-gen will not use the hashing file.

Fixes: 119117395
Test: N/A
Change-Id: I82a309b401e326ddbc58949479b386f3c57f8c1f