Add missing output of doc comments.
Some of the code that emits doc comments was only emitting
doc comments for some branches. Until b/37749857 is fixed (leaks in
AIDL/HIDL), we can't guarantee that a doc comment is always emitted.
Here, we add the known missing emissions. We also check to make sure
that a doc comment isn't emitted multiple times for sanity.

Bug: 130911129
Test: manually inspect output
Test: m $(allmod | grep -P "@[0-9]\.[0-9]$" | grep -v com.android)
    (heuristic to build all HIDL C++ interfaces)
Change-Id: Ib1ae438d50def7370640b0cd59d05dc93fb3e598