safe_union: fix uninitialized union data.
Originally, safe_union was always invalid when it was initialized.
However, now it is initialized to the first element. Because of this, we
need to memset this memory region when the type is initialized.

b/131437783 is filed as a follow-up to avoid extra memsetting (assuming
the compiler doesn't already optimize it out).

Bug: 131356202
Test: hidl_test (tests in follow-up CL)
Change-Id: I756a4062d1b921633c36b8dcca91dfc088a722f0
(cherry picked from commit d6bbd01489e64a7e2335d681f39d0f8b0bbf983d)
Merged-In: I756a4062d1b921633c36b8dcca91dfc088a722f0