Remove static gAllReservedMethods from Interface.cpp
This was a global static that was used to determine all the methods in
IBase so that they could be added to all inheriting children. This
removes that global static and adds it as a member in AST.

Bison now calls into AST to add a method to an interface, which the AST
uses to store the reserved methods (similar to how gAllReservedMethods
did). The call to addAllReservedMethods also goes through AST which
finds all the reserved methods by checking all its importedAST (one of
which is guaranteed to have the methods).

Bug: N/A
Test: ./test/run_all_host_tests.sh && ./test/run_all_device_tests.sh
Change-Id: I47c592d724047c8e78e7b978b46837b39ffe4221