hidl_package_root:detect path/use_current conflict
'path:' is hanging around for legacy reasons, although it is no longer
required to specify it (it defaults to ModuleDir), many places still do.
However, when use_current is used, ModuleDir()/current.txt is used as
the path. However, hidl-gen will always use the path as specified as
'path:', even if it is not the same as ModuleDir.

A simpler fix might be do switch use_current to tell the build system to
check for the existance of 'path:'/current.txt. However, sometimes path
is outside of the ModuleDir, and Soong modules can't reference paths
outside of their module dir. So, with this patch, we instead error when
use_current is used and 'path:' does not match ModuleDir.

Ideally, we would simply move all hidl_package_root's to their 'path:'
and then remove 'path:' from the build system, but this is cleanup work
still todo.

Fixes: 138253824
Test: try building a hidl_package_root w/ an unrelated 'path:'
Change-Id: I44e01bcea1fcf6d8f0b89fe987c7c7a507b5d061