Output lints to standard out
When interfaces fail to build they output to standard error, but on
presubmit this standard error is redirected to the lint.json file. This
makes build breakages hard to debug.

This CL moves the lints to stdout so that if there are build errors the
errors would show up on stderr.

Bug: 139094646
Test: TH
Test: Run hidl-lint and make sure there is good error output
Test: Check hidl-lint.zip file in build artifacts to make sure lints
still show up

Change-Id: Icb7437d9a4b77a2fff16b46018f50c1963a0a5f5