Only addImportedNamesGranular when success.
Before, 'hidl-gen -v -Lcheck android.hardware.authsecret@1.0' was
segfaulting. This is because the '-v' check for unreferenced types is
getting the ASTs for all imported granular types. However, authsecret
has no types.hal file. Because addImportedNamesGranular was called even
on unsuccessful imports (which is, AFAIK, only ever accepted as a
non-error when implicitly importing types), this routine was trying to
access data in a null parsed types AST.

Bug: N/A
Test: source scripts/hal-queries.sh && hidl-gen -v -Lcheck $(aosp-interfaces)
Test: ./test/run_all_host_tests.sh
Change-Id: I96f655c3168b470c1a61f9feb783c2452b5101a5